+++ spuce/filters/root_raised_cosine_imp.h
@@ -4,5 +4,5 @@
 
 #include <spuce/typedefs.h>
 namespace spuce {
-float_type root_raised_cosine_imp(float_type alpha, float_type xin, float_type rate, long num_taps);
+float_type root_raised_cosine_imp(float_type alpha, float_type xin, float_type rate, size_t num_taps);
 }  // namespace spuce
